feat: Add Instruction to delete Python files
This commit adds a shell script, 'delete_python_files.sh', that automates the deletion of Python files (files with a .py extension) within the specified directory. The script utilizes the 'find' command to identify and remove all Python files present in the current directory when executed. Additionally, instructions on how to run the script have been added to the README.md file for reference.